+++ src/Command/RecreateCacheCommand.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dbp\Relay\BaseOrganizationConnectorCampusonlineBundle\Command;
+
+use Dbp\Relay\BaseOrganizationConnectorCampusonlineBundle\Service\OrganizationProvider;
+use Symfony\Component\Console\Attribute\AsCommand;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+#[AsCommand(
+    name: 'dbp:relay:base-organization-connector-campusonline:recreate-cache',
+    description: 'Re-create the course cache',
+)]
+class RecreateCacheCommand extends Command
+{
+    public function __construct(private readonly OrganizationProvider $organizationProvider)
+    {
+        parent::__construct();
+    }
+
+    protected function execute(InputInterface $input, OutputInterface $output): int
+    {
+        $return = Command::SUCCESS;
+        try {
+            $this->organizationProvider->recreateOrganizationsCache();
+        } catch (\Throwable $exception) {
+            $output->writeln('failed to re-create organization cache: '.$exception->getMessage());
+            $return = Command::FAILURE;
+        }
+
+        return $return;
+    }
+}