+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- remove failing and unnecessary migration
+- Fix migration: Update tables (add foreign key constraints) instead of re-creating them 
+- Add re-create cache command
 
 ## v0.2.3